User can view others' PPs. Not fully integrated with Kovi due to conflicting code
Eug, I changed the hyperlink in the homepage for the users to redirect to their PPs.

Kov, can you try to link it up so that users can view PPs other than their own? I used Request.QueryString["userId"] to get the id of the user I want to view. You can refer to my codes in ProfilePage.Master.cs.