[ IMPORTANT ] Please read description
I manged to block past dates on the calendar in Create Event page, and cleared the cache each time you log out.

IMPORTANT. I have deleted a column called EventAdv or something like that, so please make the necessary changes to your DAO, BLL, etc. I hve already deleted mine on the evetnDAO and the attributes in my event BLL. AGN, please choose the left side when merging the db, AND PICK THE SAME SIDE for the log file as well.